Update 1
The stream is now working.
This update is preparation for the new functions.

Changelog: 
Var for the stream string
Two values for 'a' table: 'a' and 'b'.
New functions: 'jtrCreateSound', 'jtrManageSound' and 'jtrFailSafe'.
'jtrCreateSound' now creates the stream at LocalPlayer's position in the world.

Notes:
The stream cannot be destroyed or stop at the moment. Be careful when spawning radio in you'll have the stream on until map change, server shutdown or restart.